21 Card Game
A text-based card game that can be played interactively with a computer. It is played with 2 standard decks of cards. The goal of the game is to reach 21 points or get as close to 21 points as possible. If the computer's score is higher than yours, you lose. At the beginning of the game, each player (the computer and the user) is dealt 2 cards. Then, the user is asked whether they want to use their extra card drawing option to gain points. After the user finishes drawing cards, it's the computer's turn. After all actions are completed, the scores are calculated, and the winner is determined.

I first created this project in Java, then I reorganized and wrote it for Python language.